somersaultlazy: Convert it to the 'isotp' module
It turns out that with a few tricks this module can also be used in
conjunction with asyncio. Compared to aioisotp, we get much better
support for configuring the parameters of the in-kernel isotp
subsystem, e.g., being able to instruct it to pad the last CAN frame
of an ISO-TP telegram to 8 bytes. On the flipside, we now require a
relatively recent linux kernel (>= 5.10), and we lose the userspace
ISO-TP implementation, i.e. we now must use a socketcan interface
instead of being completely self-contained in unittest mode.

Signed-Off-By: Andreas Lauser <Andreas.Lauser@mbition.io>
Signed-Off-By: Katrin Bauer <Katrin.Bauer@mbition.io>