Fix various issues (#78)
* DOCREF: keep IDs valid python names

We now use the naming scheme f"{id}_from_{doc}". the previous scheme caused
problems with the NamedItemLists of the comparams of ecus, because

```python
ecu.comparams.FOO>BAR
```

is not a valid python expression.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>

* fix copy-and-pasto

The relevant variable is called `dr`, not `cdi`. We need better tests!

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>

* make the EnvironmentDataDescription class inherit from DopBase

also, add implementations for `convert_X_to_Y()` functions to make the
class instantiable and add a bit_length attribute which is assumed to
exist by some places.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>

* make the loading of tools more robust

now, if one tool cannot be loaded for some reason, the other ones
still can be used. The motivation for this is that on my system the
PyInquirer dependency broke recently. Even though PyInquirer is only
used by the `browse` tool, it still prevented any tool to be used...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>

* consider review comments

thanks [at]kayoub5!

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>