refactor the ID/reference mechanism for ODXLINK
instead of being a simple string, an ODXLINK ID of an object now
consists of the object's document-local ID and the corresponding
document fragment instead of just a simple string. References are the
same as IDs, but if no document is specified, it means "same document
as the referencing object".

while at it, fix mypy for the tests, examples and setup.py and as a
consequence, run it for the whole project instead of just the
`odxtools` sub-directory.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: David Holtz <david.holtz@mbition.com>