allow the IDs of communication parameters to contain hyphens
in contrast to short names, IDs are not automatically valid python
identifiers. (they are defined as exhibiting the xsd:ID type, which --
if I did not make a mistake again -- can follows the same conventions
as python identifiers but can also contain dots and hyphens after the
first character.)

thanks to [at]kayoub5 for the catch.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: David Holtz <david.holtz@mbition.com>