make sure that the items of NamedItemLists stemming from short names are identifiers
it turned out that short names are allowed to start with digits, while
python identifiers aren't. (in this case, we add an underscore as a
prefix the short name.)

besides this, this patch reduces the usage of `lambda` functions in
favour list comprehension.

finally, the communication parameter handling of DiagLayer is improved
and the correct parameter for the tester present time is
used. (hopefully)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: David Holtz <david.holtz@mbition.com>