remove the "CANdela workaround" hacks
The reason for these were that CANdela uses different communication
parameter specifications than the ones specified by the `odx-cs` files
that are shipped with the ODX standard. This resulted in a different
ordering for some of the complex communication parameters (albeit with
all required fields being present). Fortunately, with #92 merged, this
became unnecessary.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>