add support for special data groups (SDGs)
Be aware that I might have missed a place or two where SDGs are
allowed because the inheritance hierarchy defined by the ODX XSD is
not identical to the one of the python code.

Besides support for SDGs, this patch fixes various issues which I've
stumbled over while testing it. (Side note: we *need* more
comprehensive unittests!)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>