use named constructors for constructing objects from XML where it makes sense
These constructors mostly relace the `read_<foo>_from_odx()` functions
and follow the convention `<Foo>.from_et()`. This approach has the
advantages that much more of the code related to a given class is
contained within the class' body and that only the class name (`Foo`)
needs to be imported for consumers.

This patch only replaces the "constructor-like" functions, i.e., those
which return a single object of a specific class. (The remaining ones
will be dealt with by a later patch.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>