diag layers: fix determination of comparams
this got broken by a2c8e0ea: The problem with the new simplified
approach was, that comparameters with the same name that were
specified by low-priority parents were not properly overwritten by the
higher-priority ones. Instead, the latter were just appended to the
former and this obviously did not work if the first specifified
occurence of a communication parameter was used. We now revert back to
building a dictionary, but we (a) use a (comparam.short_name,
protocol_short_name) tuple as the key and (b) treat the dictionary as
an internal implementation detail of the function (i.e., we return a
list).

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>