fix handling of boolean attributes
in ODX, a true value can be specified by either the strings "true" or
"1", and false one by the strings "false" or "0". Also, leading and
trailing white space is to be ignored. All of these complications are
now dealt with by the `str_to_bool()` function.

Further, if an attribute of the XML is optional but specified to
default to true, `None` must be interpreted as true, which is done by
assigning the raw value to a protected attribute `._<foo>` while also
defining a property `foo` that returns a boolean...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>