shorten the stringification of OdxLinkId and OdxLinkRef objects
in more than 90% of all cases, the document fragments are not relevant
for the user, and the string representation of these objects is mainly
for the user's benefit.

note that the `__repr__()` function stays unchanged because python
wants its result to be encompassing, i.e., `eval(repr(obj)) == obj` is
supposed to hold.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: David Holtz <david.holtz@mbition.io>