list objects should never be optional
if a tag for a list is not present in the XML, this simply corrosponds
to an empty list on the python side. this approach has the hugh
advantage that it is not necessary to ensure that these objects are
not `None` before looping over them, and because the XML does not
allow empty lists tags, the respective tag can be skipped if the list
object is empty.

thanks to [at]kayoub5 for pointing this out.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Kevin Klein <kevin.klein@mbition.io>