More fixes (#118)
* EcuVariantMatcher: use [at]staticmethod for the `_from_et()` function

[at]staticmethod is better because the resulting code is sightly more
explicit and all other `_from_et()` functions use it as well.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Kevin Klein <kevin.klein@mbition.io>

* slightly refactor EndOfPduField

the XSD only allows to specify structure references, not structures
themselves. Also, the environment data description references were
missing...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Kevin Klein <kevin.klein@mbition.io>

* list objects should never be optional

if a tag for a list is not present in the XML, this simply corrosponds
to an empty list on the python side. this approach has the hugh
advantage that it is not necessary to ensure that these objects are
not `None` before looping over them, and because the XML does not
allow empty lists tags, the respective tag can be skipped if the list
object is empty.

thanks to [at]kayoub5 for pointing this out.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Kevin Klein <kevin.klein@mbition.io>

* DiagDataDictionarySpec: it expects NamedItemLists directly now

i.e., no need to convert normal lists to them in post-initialization
anymore.

thanks to [at]kayoub5 for noticing this!

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Kevin Klein <kevin.klein@mbition.io>

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Kevin Klein <kevin.klein@mbition.io>