refactor the initialization and reference resolution mechanism
Now, all non-primitive objects which are part of a diagnostic layer
provide a `._build_odxlinks()` method to create a dictionary of all
`OdxLinkId` mappings contained in this object. These dictionaries are
then collected for the whole database and used to resolve ODXLINK
references in the `._resolve_odxlinks(odxlinks)` method. Once this is
done, the diagnostic layers deal with inheritance of objects,
whereafter the short name references (SNREFs) are resolved via the
`._resolve_snrefs(diag_layer)` methods.

The initialization of objects which are not part of any `DiagLayer` is
slightly different: The `Database` object has a `refresh()` method
which deals with the full initializtion logic of all objects it
contains, the `DiagLayerContainer` and `DiagLayer` classes have
`_build_odxlinks()` and `_finalize_init()` methods. `_finalize_init()`
deals with inheritance and SNREF resolution for all objects contained.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>