diaglayer: refactor the communication parameter handling
it turns out that the names of the communication parameters are not
unambiguous: e.g., "CP_UniqueRespIdTable" parameter is not so unique
in the sense that it exists for CAN (comparam subset "ISO_15765_2") as
well as for DoIP (CP subset "ISO_13400_2_DIS_2015") and both
parameters have a completely different definition. (The same issue
applies to e.g., "CP_TesterPresentTime", but this being a simple
integer parameter, at least the returned value is of the same type.)

To get out of that pickle, the full parameter ID now has to be
specified for `DiagLayer.get_communication_parameter()` (e.g.,
"ISO_15765_2.CP_UniqueRespIdTable").

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>