DiagLayer: revert back to using short names (+protocol) to retrieve the comparams
this is rather unfortunate because we have to rely on the stars to
align correctly (only a single comparam of a given name must be
defined for an individual (protocol, is_functional) combination) to
retrieve the correct comparam. Unfortunately the IDs of comparams (and
the comparam subsets?) do not seem to be sufficiently standardized
accross the vendors to rely on them (CANdela and the comparams that
are shipped with the standard us the convention
<COMPARAM_SUBSET>.<COMPARAM_NAME> as their convention for IDs, other
editors seem to use random strings)...

thanks to [at]kayoub5 for pointing this out.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>