DiagLayer: create a DataDictionarySpec object in the logical view
This puts the sub-objects of DATA-DICTIONARY-SPEC in the same place as
where they can be found in the XML and the raw diaglayer. e.g., the
data object properties are contained by `.data_dictionary_spec`
instead of being directly helt by the DiagLayer. (Whether using
DIAG-DATA-DICTIONARY-SPEC makes much sense in the first place is a
different matter IMO...)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>