DiagLayer: simplify the `str()` and `repr()` methods
`repr()` was quite incorrect, as as this method is intended to return
a full representation of the object as a string (e.g. how an
equivalent object can be constructed), which is quite unpractical
because for `DiagLayer` this would often result in a string with a
length of several million of characters.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>