add decode support for global negative responses.
this is surprisingly complicated...

Note that, in oder to be able to decode these,
`DiagLayer.decode_response()` must be used, *not*
`DiagService.decode_message()`. This is because the diagnostic service
object does not know the diagnostic layers where it is contained
within and -- even if it knew -- it is impossible to know which diag
layer is the "correct" one.

Further, for the same reason as decoding cannot be done via the diag
service, encoding of global negative responses works via the
respective `Response` object: This object can be obtained from the
`DiagLayer.global_neg_responses.<response short name>`...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Pascal Hirmer <pascal.hirmer@mbition.io>