somersault ECU: reorder the definition of the data
this moves the definition of the tables, multiplexer, environment data
and environment data descriptions behind the definition of the global
negative responses. This is required because in a later commit, we
would like to let the `structure_ref` of the table objects to actual
structures, which is greatly simplified if the structures are already
defined before the tables.

This commit does not change anything functional in these objects, but
since git does not recognize that the code simply moved en-bloc, it
makes it very hard to review this if there were any functional changes
mixed in.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>