implement implicit determination of TABLE-KEY parameters
i.e., for encoding table-struct parameters, one now needs to specify a
dictionary with the short name of the selected table row that maps to
the value of the row's structure or DOP. If there are any conflicts
determining the row (e.g., the if the value of the TABLE-KEY parameter
and the selected row of the corresponding TABLE-STRUCT disagree), an
`EncodeError` is thrown.

This scheme was proposed by [at]kayoub5. Thanks!

Besides this, the mechanism for implicit determination of lenth keys
was simplified: we now store the value in
`encode_state.parameter_value[length_key_short_name]` instead of in a
separate `encode_state.length_keys` attribute.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>