remove `__repr__()` and `__str__()` from the parameter classes
since these are now data classes, these methods are automatically
generated and the manual variants are thus no longer necessary. Since
the these auto-generated methods are more comprehensive than the
manual ones, this implied a few minor correction in the writing code
and the unit tests to keep my personal test script working.

thanks to [at]kayoub5 for noticing this.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>