ParameterWithDop: ensure that the `.dop` attribute always exists
before this, accessing the `.dop` attribute of such parameters could
have caused an `AttributeError` to be caused if no DOP reference was
given. note that even after this patch currently it will be `None` in
this case.

besides this, the `.dop` attribute of `InputParameter` is renamed to
`.dop_base` to reflect the XML tag which produces this property
(`DOP-BASE-REF`).

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katja Köhler <katja.koehler@mbition.io>