remove the `ProgCodeSyntax` enum class
instead, make this a string because the SYNTAX subtag of
SINGLE-ECU-JOB is specified to be a plain string in the ODX
specification. It seems like we wanted to be more strict on our own
files and tried to constrain the possible values for this. This was
implemented using a string literal (which is like a type-checking only
enum, i.e., it does not check if the value specified is valid if can
only be determined at runtime). After changing this to an `Enum` class
in #172, it got some (incorrect) teeth which bit users that used
non-Java single ECU jobs.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katja Köhler <katja.koehler@mbition.io>