rework .is_required and .is_optional methods of parameters
`Parameter.is_optional()` and `Parameter.is_required()` were quite
confusing as it can be assumed that one is just the negation of the
other. We thus rename `.is_optional` to `.is_settable` to express that
a parameter is user-definable. (In the process, a few instances in the
`browse` tool where this seems to have been mixed up are fixed.) While
at it, make these functions properties so that they do not need to be
called anymore...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Pascal Hirmer <pascal.hirmer@mbition.io>