remove `DataObjectProperty.get_valid_physical_values()`
this is replaced by the texttable's `get_scales()`
method. (`get_valid_physical_values()` does not make sense for other
DOPs.) Besides the rename and the way to access it, the method returns
a list of `CompuScale` objects, so to get a list of the physical
values, the `.compu_const` property must be extracted from each of
them.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>