unfreeze `OdxLinkRef` and derive `RelatedDiagCommRef` from it
the main use case for frozen dataclasses is that they are
hashable. Since this is not a concern for `OdxLinkRef` objects, let's
unfreeze it and allow derived classes.

Besides `RelatedDiagCommRef`, there is also `ComparamRef` which should
be derived from `OdxLinkRef`, but this is more problematic because
`ComparamRef` contains quite a few attributes which the user is
usually interested in that are not in the `Comparam` specification
object (e.g., the actual value) of the communication parameter.

As usual, thanks [at]kayoub5 for the nudge.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>