slightly improve writing the sub-tags stemming from ELEMENT-ID
The jina macro for writing the SHORT-NAME, LONG-NAME and DESC tags is
now called `printElementIdSubtags` to avoid giving the impression that
it writes an `ELEMENT-ID` tag instead of just its contents.

Unless a better scheme comes along, this will also be the new way to
deal with inherited classes in the writing code: If Bar derives from
Foo, there will be `print{Foo}Subtags` and `print{Foo}Attributes`
macros, which are then called by `print{Bar}`. The first two macros do
not write the complete tag, but only the contents and the attributes
defined by Foo.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>