DiagLayer: change the way of determining whether to use CAN-FD
we now check if the string `CANFD` occurs in the value of the
`CP_CANFDTxMaxDataLength` communication parameter. if not, or if the
communication parameter has not been specified, CAN-FD is not used.

(this seems to be a bit of a kludge, but the description of
`CP_CANFDTxMaxDataLength` parameter implies that this is indeed the
way that usage of CAN-FD is specified in ODX...)

thanks to [at]kayoub5 for his help in figuring this out.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Florian Jost <florian.jost@mbition.io>