DiagCodedType: use `bitstruct.c` if available
this should come with a sizeable speedup. the main disadvantage is
that `bitstruct.c` only supports `bytes` objects but no strings,
i.e. we have to do the de- and encoding ourselfs. (Note that we want
to do this anyway because the current error checking code was
incorrect for UTF-8 strings containing multi-byte characters.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katja Köhler <katja.koehler@mbition.io>