remove the `--plumbing-output` arguments
for `list` the new `--dump-database` argument fills the gap (more or
less), for the other commands it is IMO not necessary because in their
present form, I cannot see their usefulness for non-interactive
scripts.

the `list` command's new `--dump-database` option is useful for
debugging purposes: it is about an order of magnitude faster than
`odxtools list -a $PDX_FILE`, and -- as a bonus -- it prints all
information that is internalized by the database object.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>