snoop: correctly detect response types and streamline terminal output
Once upon a time, the `response_type` property of `Response` objects
used to be a string. In modern times it has become an enum in order to
make it play nicely with type checking.

Besides this, the output of the snoop tool is slightly decluttered,
i.e., it does not prefix everything which is send by the tester by
`tester:` and everything send by the ECU by `ECU:`. (this is IMO quite
clear from the context.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>