refactor `DecodeState`
the purpose of this is to make the decoding routine a bit more
transparent and slightly more performant:

instead of copying the blob to be decoded, we now keep this constant
and have an `origin_position` attribute besides `cursor_position`. The
`origin_position` is the position within the blob to be decoded to
which relative positions of positionable parameters are ancored
to. This allows objects like structures, to just modify the origin
position of the decode state instead of creating a completely new one.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katja Köhler <katja.koehler@mbition.io>