diag coded types: refactor the decoding methods
these methods are now called `decode_from_pdu()` and are the template
for all other objects which can be decoded (i.e., DOPs and
parameters): they take a `DecodeState` object as their lone argument
and return the decoded physical value. Note that these methods are
supposed to change the passed `DecodeState` object so that it is set
up to decode the next object after calling `decode_from_pdu()`.

This reduces the mental load required because before this, parameters,
diag coded types, and DOPs all used the same name for functions which
do similar but slightly different things...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>