DecodeState: make `cursor_bit_position` an integer
i.e., it is now always an integer, never `None`. This looses us the
ability of distinguishing between "no bit position defined" and "bit
position set to 0" in the decoding machinery, but this is not too
relevant as the ODX spec says that, the bit position ought to be
assumed 0 if it is not defined and policing if it is allowed to be
defined or not ought not should not be part of the decoding logic.

thanks to [at]kayoub5 for his persistence!

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>