parameters.decode_from_pdu(): only return the value
this used to be a (value, cursor_position) tuple. With this all
decode_from_pdu() functions behave the same: they take a DecodeState
object and return the value which corresponds to this object. Also,
the DecodeState passed is updated internally to the location where the
next item to be decoded is supposed to be.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Florian Jost <florian.jost@mbition.io>