refactor limits
this is surprisingly difficult: the problem is that the LIMIT tag does
not specify the data type (integer, float, string, bytefield) which
the limit is about but the limit obviously needs to know that type in
order to do something. The solution taken here is to pass the type of
the limit's value to the `from_et()` method as an extra argument. For
Multiplexer cases where even the type of the limit's value is not
known when the limit must be created, (this is only known after the
switch key reference has been resolved) the method
`Limit.set_value_type()` is added.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Michael Hahn <michael.hahn@mbition.io>