improve the free parameter info messages
With this the `parameter_info()` functions handles more parameter
types, deals with more DOPs type and handles complex DOPs (structures)
properly. Also, I tried to make the results more consistent, but IMO
this is not too important here, because the result of this function is
supposed to be informal and for human consumption.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Michael Hahn <michael.hahn@mbition.io>