EncodeState: default the position attributes to 0
While in theory it is possible that a partially done encoded PDU ought
to be continued (which requires creating EncodeState objects with
positions other than 0), this is rare (and in fact, does not occur
inside odxtools yet), so these attributes should be defaulted to 0.

as usual, thanks to [at]kayoub5 for the catch.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>