improve detection of overlapping parameters during encoding
this code should now be bullet-proof: instead of complaing only if at
least of the bits which are supposed to be set by a parameter in the
coded message is already set, used bits are now explicitly tracked via
a bit mask in the `EncodeState`. As a result, if e.g. an all-zero
integer parameter overrides another all-zero integer parameer we will
now get an overlapping objects warning as we are supposed to.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Alexander Walz <alexander.walz@mbition.io>