fields: fix end-of-PDU handling
an element of a field is at the end of a PDU if the field is at the
end of the PDU and the item is the last element of the field. Note
that even though some pathetic cases where this makes a difference can
be constructed, I have doubts if this matters in practice.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>