Improve the file loading mechansims
this refactors the file loading mechanism with the goal to create a
database from multiple constituive files. It works by introducing a
new user facing `odxtools.load_files(filename1, filename2, ...)`
function.

Besides this, `odxtools.load_directory(dirname)` is added. This
function loads all files contained in a directory into a single
database. The main application is probably to directly load an
unzipped PDX file. This is a quite convenient feature when developing
tools or a dataset.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Florian Jost <florian.jost@mbition.io>