make descriptions conform to the specification
even though I never saw DESC tags containing anything other than a
subset of HTML, the specification allows them to provide a list of
URLs to auxiliary external documents and a text identifier (the TI
attribute).

To make working with these objects as easy as possible and in order to
simplify backwards compatibility, the `__str__()` method of these
`Description` objects returns only the content of the tag.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mbition.io>