change interpretation of "last parameter"
the specification states that if the position of a parameter is not
explicitly specified using BYTE-POSITION, "The parameter starts then
at the byte edge next to the end of the last parameter". This patch
changes the interpretation of "last parameter" from "parameter
currently furthest to the end of the PDU" to "most recently handled
parameter".

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katja Köhler <katja.koehler@mbition.io>