handle auxiliary files more comprehensively
this means writing the auxiliary files contained in the database as
well as making the data of auxiliary files referenced by `ProgCode`
objects conveniently available. This is surprisingly difficult because
it implies some modifications to how a few objects are deepcopied in
order to avoid coping the whole database when just copying a
diagnostic layer.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>