do not greedily load the contents of auxiliary files
as [at]kayoub5 rightfully noted, some types of ODX files (usually
`.odx-f`) may feature auxiliary files that might be several Gigabytes
in size. What we do instead is to store file handles to these
files. (`ZipFile` thankfully provides that without having to
uncompress the whole archive.) This approach comes with the
disadvantage that these files must not be modified on disk while they
are used by odxtools, but IMO this is trade-off is worthwhile.

thanks to [at]kayoub5 for his insistence.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>