do not keep commented out code in the repository
the reason why this was commented out is that the object goes out of
scope immediately afterward, so it is unnecessary to change it back to
`None`. the reason why the line was kept in the first place is that it
is conceptually what is supposed happen. Anyway, since this is in no
way performance critical, let's just add the line back.

thanks to [at]kayoub5 for the catch.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mbition.io>