properly support decoding NRC-CONST parameters
if a negative response contains an NRC-CONST parameter, where the
parameter's value is not in the list of expected parameters, the
response is not applicable and decoding should be stopped at this
point. (Thus, NRC-CONST parameters are a multiplexer mechanism which
filters response objects based on the value of some other parameter.)

This patch implements this by introducing a new `DecodeMismatch`
exception which is raised whenever a NRC-CONST parameter which
exhibits an unexpected value is encountered. The higher level decoding
routines for decoding responses in `DiagLayer` then ignore these
exceptions. (IMO, this solution is a bit sub-optimal because it uses
exceptions for flow control in normal operation, but alternative
implmentations would be far more complex.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katja Köhler <katja.koehler@mbition.io>