implement diagnostic variables
In true "design by comitee" spirit, the ODX specification supports to
layer an entirely different communication paradigm over the default
RPC-based model: Diagnostic variables. In this model, communication
with the ECU is realized by reading from and writing to a set of
"variables". ODX glues this communication approach on its "normal"
service-oriented model by specifying a set of variables plus setter
and getter services for each.

Note that odxtools can only expose the required information for this,
not implement this approach itself, because odxtools is not concerned
with communicating with ECUs directly.

Besides communication variables, this patch also adds the code to
interalize the information necessary for dynamically defined messages
(`DYN-DEFINED-SPEC`), i.e., messages which exhibit formats that are
defined at runtime instead of being known a-priori.

Note that, since I don't have any data file which use either of these
features, your milage may vary.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Florian Jost <florian.jost@mbition.io>