SDGs: if the SDG caption is referenced, do not add its IDs to the odxlink database
This prevents these SDG captions to be added multiple times (i.e., for
SDGs which only reference the caption, not define it). Under normal
circumstances, this only results in slightly lower performance, but in
special situations like having to modify a database based on another,
it can lead to SDG captions added which are not defined. (I recently
stumbled over such a case.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Florian Jost <florian.jost@mbition.io>