fix the names of types for transfer functions
Since `COMPU-PHYS-TO-INTERNAL` and `COMPU-INTERNAL-TO-PHYS` are
conceptually inverse functions of each other, the set objects which
they are mapping to each other are swapped. i.e., the "source" type of
`COMPU-PHYS-TO-INTERNAL` is the physical type of the compu method
while `COMPU-INTERNAL-TO-PHYS` it is the opposite. Let's avoid this
confusion and use the proper mathematical terms 'domain' (for the set
of objects which a function maps from) and 'range' (for the set of
objects which a function maps to). Note that the domain and the range
are swapped for `COMPU-PHYS-TO-INTERNAL` and `COMPU-INTERNALTO-PHYS`.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Michael Hahn <michael.hahn@mbition.io>