introduce OID attributes
OID stands for "Object ID" and is meant to be a UUID for objects,
i.e. it is supposed to be an invariant globally unique identifier for
objects. It can e.g., be used to locate objects in an ODX editor
across projects, but there is no formal mechanism to reference an
object via its OID within the ODX data model.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>