fix style of last argument of argument lists
instead of sometimes using a comma for the last item of a list and
sometimes not, we now always close the list in the same line. This
does not come with any functional changes, but it makes the coding
style a bit more consistent and IMO the code is slightly more readable
with this style.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>