Implement the `RAT-FUNC` compu method
This compu method uses transfer functions based on a rational
functions. Note that, since rational functions are not invertible in
general, the physical-to-internal mapping is only available if it is
explicitly specified via COMPU-PHYS-TO-INTERNAL and thus parameters
which use such compu methods can only be encoded if both directions
are specified. (TODO: Add possibility to directly specify internal
values for the purpose of encoding messages.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mbition.io>