fix the document fragments for objects not featured by the base diag layer
since diag layers constitute a document fragment in addition to the
container which contains them, the odxlink ids must reflect that. So
far, this did not happen for objects which are added by one of the
concrete diagnostic layers classes, i.e. objects which were not
handled by the diagnostic layer base class.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mbition.io>