Add codec objects
Codecs are any objects that can be en- and decoded into PDUs. The
approach so far was to derive all "composite" codecs from
`BasicStructure` and to use a subset of that API for "simple" codable
objects like DOPs or parameters. In particular, the previous approach
to composite objects was sub-optimal because `BASIC-STRUCTURE`
features subtags that other codable objects like `Request` and
`Response` do not exhibit (i.e., `.byte_size`).

Since codecs do not play very nicely with inheritance, the new
approach is based on typing protocols (Java would call these
"interfaces"): `codec.py` defines a few runtime checkable
`typing.Protocol` classes which codecs must implement: `Codec` for the
basic functionality, `CompositeCodec` defines the API for codecs that
are composed of a list of parameters like structures, requests,
responses or environment datas, and `ToplevelCodec` defines the
"external" API for requests and responses that is supposed to be
called by user scripts. Any codable class implementing these APIs can
check if it is compliant using `isinstance()` in the `__post_init__()`
method of the respective class.

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Alexander Walz <alexander.walz@mercedes-benz.com>
Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>