TableKeyParameter: do not retrieve table object from potentially unresolved reference
When resolving the table row in `TableKeyParameter`, whether or not
the references of that object have been resolved is undefined. Since
the table is a reference in table rows, so far an exception gets
raised in `TableKeyParameter` depending on the phase of the
moon. (i.e., whether `._resolve_odxlinks()` has been called at this
point for the table row in question or not.)

This has been discovered by @mafei1982. Thanks!

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mercedes-benz.com>