`retarget_snrefs()`: assign the database and the diaglayer independently
While it does not make much sense to do so (the idea is if the
`SnRefContext` is created internally in the toplevel call to the
function, the `.database` and `.diaglayer` attributes are assigned
automatically), doing it separately makes the code slightly easier to
understand.

thanks to [at]kayoub5 for the catch!

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>