improve the fallback code for some undefined encodings in non-strict mode
For A_INT32, we now allow the BCD encodings (but use the absolute of
the provided value), for A_UINT32 we now allow signed encodings and
simply pass through the raw number. Also, we now complain if the value
that shall be encoded does not fit into the specified bit length.

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mercedes-benz.com>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mercedes-benz.com>