refactor the ECU variant matching code
On its own, this refactoring does not make much sense because it adds
abstractions without adding a second implementation for it. This
will change in the next commit which adds base variant patterns for
identifying the base variant.

This commit also contains quite a few cleanups and fixes some smaller
issues with the ECU variant matching code.

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Christian Hackenbeck <christian.hackenbeck@mercedes-benz.com>