Implement state machines
This implements the runtime functionality of state charts. Note that
the specification strongly implies deterministic state machines, but
does not explicitly forbid non-determinism and also does not specify
what should happen if more than one state transition ought to be
taken.

Also, IMO it is rather inconvenient that no values for state
transition references which specify constant parameters can be
specified because the parameters relevant for the state transitions
are (probably, the spec is also unclear here) the ones of the
response, and these can vary depending on the concrete data
received. Finally, it is IMO rather limiting that value parameters can
only be tested for equivalence, and not for scale-constraints...

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mercedes-benz.com>