StateTransitionRef: Also check if it applies to the request
I think contitions for state transitions make only make sense for the
received responses, but the spec is ambiguous here and as [at]kayoub5
rightfully notes, also checking the request most will not do anything
except in pathologic cases.

note that this change implied the introduction of a `lenient` argument
for `resolve_snrefs()`, because we need to be able to check if a
parameter via SNREF is present in strict as well as in non-strict
mode.

thanks to [at]kayoub5 for the nudge.

Signed-off-by: Andreas Lauser <andreas.lauser@mercedes-benz.com>